+++ client/webpack/helpers/environment.js
@@ -1,5 +1,5 @@
 import path from "path";
-import dotenv from "dotenv"
+import dotenv from "dotenv";
 
 dotenv.config({ path: path.join(__dirname, "../../../.env") });
