--- client/webpack/servers/webpack-dev-server.js
@@ -1,53 +1,0 @@
-import environment from "../helpers/environment";
-import webpack from "webpack";
-import WebpackDevServer from "webpack-dev-server";
-import webpackConfig from "../config/browser-dev.config";
-import ch from "../../src/helpers/consoleHelpers";
-import paths from "../helpers/paths";
-
-const compiler = webpack(webpackConfig);
-let timer;
-
-compiler.hooks.done.tap("ManifoldWebpackDevServer", stats => {
-  const end = +new Date();
-  const ms = end - timer;
-  setTimeout(() => {
-    ch.background(`Browser assets have finished building (${ms}ms).`);
-    ch.header(
-      `Browser assets available at: ${environment.devUrl}/webpack-dev-server`
-    );
-  }, 250);
-});
-
-compiler.hooks.compile.tap("ManifoldWebpackDevServer", params => {
-  timer = +new Date();
-  ch.background("Serving browser assets from webpack-dev-server.");
-});
-
-const hot = !process.env.DISABLE_HMR;
-const allowedHosts = [
-  "manifold.lvh",
-  "localhost",
-  "127.0.0.1",
-  "manifold-dev.lvh",
-  "manifold-stable.lvh",
-  "manifold-dev.ngrok.io"
-];
-if (process.env.DOMAIN) allowedHosts.push(process.env.DOMAIN);
-
-const serverOptions = {
-  hot,
-  host: process.env.DOMAIN || "localhost",
-  allowedHosts,
-  port: environment.devPort,
-  headers: { "Access-Control-Allow-Origin": "*" },
-  static: {
-    directory: paths.root
-  }
-};
-
-const server = new WebpackDevServer(serverOptions, compiler);
-
-(async () => {
-  await server.start();
-})();
